EOS-14134: move SWUpdate functionality to seperate file to avoid cross-import
